wyglad kodu

+++ GUI/Simulation2/NotEdited.cpp
@@ -48,13 +48,21 @@ void advect(int N, int b, float* d, float* d0, float* u, float* v, float dt)
 {
 	int i, j, i0, j0, i1, j1;
 	float x, y, s0, t0, s1, t1, dt0;
-	dt0 = dt * N;
+	dt0 = dt * (float)N;
 	for (i = 1; i <= N; i++) {
 		for (j = 1; j <= N; j++) {
-			x = i - dt0 * u[IX(i, j)]; y = j - dt0 * v[IX(i, j)];
-			if (x < 0.5) x = 0.5f; if (x > N + 0.5) x = N + 0.5f; i0 = (int)x; i1 = i0 + 1;
-			if (y < 0.5) y = 0.5f; if (y > N + 0.5) y = N + 0.5f; j0 = (int)y; j1 = j0 + 1;
-			s1 = x - i0; s0 = 1 - s1; t1 = y - j0; t0 = 1 - t1;
+			x = (float)i - dt0 * u[IX(i, j)]; 
+			y = (float)j - dt0 * v[IX(i, j)];
+			if (x < 0.5) x = 0.5f; if (x > N + 0.5) x = N + 0.5f; 
+			i0 = (int)x; 
+			i1 = i0 + 1;
+			if (y < 0.5) y = 0.5f; if (y > N + 0.5) y = N + 0.5f;
+			j0 = (int)y;
+			j1 = j0 + 1;
+			s1 = x - i0;
+			s0 = 1 - s1;
+			t1 = y - j0;
+			t0 = 1 - t1;
 			d[IX(i, j)] = s0 * (t0 * d0[IX(i0, j0)] + t1 * d0[IX(i0, j1)]) +
 				s1 * (t0 * d0[IX(i1, j0)] + t1 * d0[IX(i1, j1)]);
 		}
@@ -107,7 +115,8 @@ void vel_step(int N, float* u, float* v, float* u0, float* v0, float visc, float
 	SWAP(v0, v); diffuse(N, 2, v, v0, visc, dt);
 	project(N, u, v, u0, v0);
 	SWAP(u0, u); SWAP(v0, v);
-	advect(N, 1, u, u0, u0, v0, dt); advect(N, 2, v, v0, u0, v0, dt);
+	advect(N, 1, u, u0, u0, v0, dt); 
+	advect(N, 2, v, v0, u0, v0, dt);
 	project(N, u, v, u0, v0);
 }
